@@ -1,17 +1,20 @@
 {
-    "name": "Wordpress Plugin Base Generator",
-    "description": "wordpress-plugin-base-generator",
+    "name": "clebermind/wordpress-plugin-base-generator",
+    "description": "Wordpress Plugin Base Generator",
+    "license": "MIT",
     "authors": [
         {
             "name": "Cleber Mendes",
-            "email": "clebermind@gmail.com"
+            "email": "clebermind@gmail.com",
+            "homepage": "https://github.com/clebermind",
+            "role": "Developer"
         }
     ],
     "require": {
-        "php": ">=5.6",
-        "symfony/console": "^4.3",
-        "cocur/slugify": "^3.2",
-        "symfony/process": "^4.3"
+        "php": ">=7.4",
+        "symfony/console": "^5.3",
+        "cocur/slugify": "^4.0",
+        "symfony/process": "^5.3"
     },
     "config": {
         "vendor-dir": "lib/vendor"
